Add a new folder GameMain, move files, delete folder GUI
