+++ src/Makefile
@@ -1,9 +1,11 @@
 all:
 	cd SequenceSimulation; make all
 	cd visualization; make all
-	cd alignment; make
+	cd alignment; make all
+	cd GameMain; make all
 
 clean:
 	cd SequenceSimulation; make clean
 	cd visualization; make clean
 	cd alignment; make clean
+	cd GameMain; make clean
