+++ src/visualization/Makefile
@@ -1,7 +1,5 @@
 all:
-	cd GUI; make all
 	cd StaticGraph; make all
 
 clean:
-	cd GUI; make clean
 	cd StaticGraph; make clean
